requirements.txt
matplotlib>=2.2
networkx>=1.11
numpy>=1.12
pandas>=0.19.2
scikit-learn>=0.18
scipy>=0.18.1
tabulate>=0.7.5
tensorflow>=1.0.1
tqdm>=4.11.2
h5py>=2.6.0
climin>=0.1a1
xmltodict>=0.10.2
keras>=2.0.2
psycopg2>=2.7.3.2
dask>=0.18.2
xarray>=0.10.9
zarr>=2.2.0
numcodecs>=0.6.1
toolz>=0.9.0
cytoolz>=0.9.0.1
msgpack-python>=0.5.6
scikit-image>=0.14.0
GitPython>=2.1.11
